backend do mapview

+++ Code/Cell.cpp
@@ -3,8 +3,8 @@
 namespace ZPR {
 	Cell::Cell(int row, int column) : _row(row), _column(column) {}
 
-	sf::Vector2f Cell::GetPosition()
+	sf::Vector2i Cell::GetPosition()
 	{
-		return sf::Vector2f(this->_row, this->_column);
+		return sf::Vector2i(this->_row, this->_column);
 	}
 }
